[gaim-migrate @ 14533] Fix code as per a warning.
--HG--
extra : convert_revision : 4dd7fdecd196d8491dd19b53337c917a657d61f8

+++ src/gtkdebug.c
@@ -347,7 +347,7 @@ regex_match(DebugWindow *win, const gchar *text) {
 		 * if necessary...
 		 */
 		do {
-			gint m;
+			size_t m;
 
 			for(m = 0; m < n_matches; m++) {
 				GtkTextIter ms, me;
